refactor(grid): isolate controller tracking
