+++ ShipCoreFramework/src/Data/Scripts/ShipCoreFramework/Shared/Components/GridComponent.cs
@@ -4,7 +4,6 @@ using System.Linq;
 using Sandbox.Game.Entities;
 using VRage.Game.ModAPI;
 using IngameIMyEntity = VRage.Game.ModAPI.Ingame.IMyEntity;
-using IMyShipController = Sandbox.ModAPI.IMyShipController;
 
 namespace ShipCoreFramework
 {
@@ -13,9 +12,7 @@ namespace ShipCoreFramework
         internal MyCubeGrid Grid;
         private IMyGridGroupData _groupData;
         private readonly object _blocksLock = new object();
-        private readonly object _shipControllersLock = new object();
         private readonly List<IMySlimBlock> _blocks = new List<IMySlimBlock>();
-        private readonly List<IMyShipController> _shipControllers = new List<IMyShipController>();
         internal int BlockCount
         {
             get
@@ -73,33 +70,6 @@ namespace ShipCoreFramework
 
             var otherBlocks = blocks.Where(block => !Utils.IsCoreBlock(block)).ToList();
             foreach (var otherBlock in otherBlocks) BlockAddedInternal(otherBlock);
-        }
-
-        internal List<IMyShipController> GetShipControllersCopy()
-        {
-            lock (_shipControllersLock)
-                return new List<IMyShipController>(_shipControllers);
-        }
-
-        private void TrackShipController(IMyShipController shipController)
-        {
-            if (shipController == null) return;
-
-            lock (_shipControllersLock)
-            {
-                if (!_shipControllers.Contains(shipController))
-                    _shipControllers.Add(shipController);
-            }
-        }
-
-        private void UntrackShipController(IMyShipController shipController)
-        {
-            if (shipController == null) return;
-
-            lock (_shipControllersLock)
-            {
-                _shipControllers.Remove(shipController);
-            }
         }
 
         private void GridMarkedForClose(IngameIMyEntity entity)
